summariser/requirements.txt
praw
google-cloud-storage
google-cloud-secret-manager
google-cloud-aiplatform
Flask